split users and group models
